requirements.txt
# asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
# Django==4.1.7
# django-jazzmin==2.6.0
Faker==18.3.1
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
razorpay==1.3.0
requests==2.28.2
six==1.16.0
# sqlparse==0.4.3
# tzdata==2022.7
urllib3==1.26.15

# ------------------------------From account----------------------------------
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3